Move report UT to correct package, add OpenApiGeneration test, move tests to controller package
