--- src/test/functional/com/MyApp/budgetControl/api/SwaggerDocsGenerationTest.java
@@ -1,31 +1,0 @@
-//package com.MyApp.budgetControl.api;
-//
-//import static org.hamcrest.Matchers.containsString;
-//import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-//import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
-//import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-//
-//import org.junit.jupiter.api.Test;
-//import org.junit.jupiter.api.TestInstance;
-//import org.springframework.beans.factory.annotation.Autowired;
-//import org.springframework.boot.test.context.SpringBootTest;
-//import org.springframework.boot.webmvc.test.autoconfigure.AutoConfigureMockMvc;
-//import org.springframework.test.context.ActiveProfiles;
-//import org.springframework.test.web.servlet.MockMvc;
-//
-//@SpringBootTest
-//@AutoConfigureMockMvc
-//@ActiveProfiles(value = "test")
-//@TestInstance(TestInstance.Lifecycle.PER_CLASS)
-//class SwaggerDocsGenerationTest extends TestContainersConfigurationSpec {
-//
-//  @Autowired
-//  private MockMvc mockMvc;
-//
-//  @Test
-//  void shouldExposeSwaggerDocs() throws Exception {
-//    mockMvc.perform(get("/v3/api-docs"))
-//        .andExpect(status().isOk())
-//        .andExpect(content().string(containsString("\"OpenAPI definition\"")));
-//  }
-//}
